swagger模板

+++ gen_code_v2/code_ui/src/main/java/org/alvin/code/v2/sys/mock/JsMockBus.java
@@ -65,12 +65,17 @@ public class JsMockBus implements InitializingBean {
 		System.gc();
 		File dirFile = new File(dir);
 		//只有两层
-		for (File f : dirFile.listFiles()) {
-			for (File subF : f.listFiles()) {
-				subF.delete();
+		if(dirFile.listFiles() != null){
+			for (File f : dirFile.listFiles()) {
+				if(f.listFiles() != null){
+					for (File subF : f.listFiles()) {
+						subF.delete();
+					}
+				}
+				f.delete();
 			}
-			f.delete();
 		}
+
 		dirFile.delete();
 		return file;
